// lib/api/generated/models/apiKeyResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Realillust API
 * OpenAPI spec version: 0.1.0
 */
import type { ApiKeyCreatedBy } from "./apiKeyCreatedBy";
import type { ApiKeyStatus } from "./apiKeyStatus";

export interface ApiKeyResponse {
  id: string;
  name: string;
  maskedKey: string;
  status: ApiKeyStatus;
  createdAt: string;
  /** @nullable */
  lastUsedAt: string | null;
  createdBy: ApiKeyCreatedBy;
}